[feat] CORS 설정 변경
